Move default implementation of DestinationRendersRegistry to the library.
