+++ composeApp/src/commonMain/kotlin/com.macaosoftware.component.navigationcompose.demo/di/ServerUiModule.kt
@@ -1,7 +1,9 @@
 package com.macaosoftware.component.navigationcompose.demo.di
 
 import com.macaosoftware.component.core.DestinationRendersRegistry
-import com.macaosoftware.component.navigationcompose.demo.marketplace.DemoDestinationRendersRegistry
+import com.macaosoftware.component.core.MacaoDestinationRendersRegistry
+import com.macaosoftware.component.navigationcompose.demo.marketplace.notfound.DestinationRenderNotFound
+import com.macaosoftware.component.navigationcompose.demo.marketplace.notfound.RootDestinationRenderNotFound
 import com.macaosoftware.component.navigationcompose.demo.serverui.domain.usecase.LoadChildrenDestinationUseCase
 import org.koin.dsl.module
 
@@ -10,6 +12,9 @@ internal val serverUiModule = module {
     factory<LoadChildrenDestinationUseCase> { LoadChildrenDestinationUseCase() }
 
     single<DestinationRendersRegistry> {
-        DemoDestinationRendersRegistry()
+        MacaoDestinationRendersRegistry(
+            destinationRenderNotFound = DestinationRenderNotFound(),
+            rootDestinationRenderNotFound = RootDestinationRenderNotFound(),
+        )
     }
 }
