+++ composeApp/src/commonMain/kotlin/com.macaosoftware.component.navigationcompose.demo/marketplace/notfound/DestinationRenderNotFound.kt
@@ -1,4 +1,4 @@
-package com.macaosoftware.component.navigationcompose.demo.serverui.domain.error
+package com.macaosoftware.component.navigationcompose.demo.marketplace.notfound
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavBackStackEntry
@@ -6,7 +6,7 @@ import androidx.navigation.NavHostController
 import com.macaosoftware.component.core.DestinationInfo
 import com.macaosoftware.component.core.DestinationRender
 import com.macaosoftware.component.navigationcompose.demo.serverui.data.ServerUiConstants
-import com.macaosoftware.component.util.DestinationRenderNotFoundView
+import com.macaosoftware.component.util.MacaoDestinationRenderNotFoundView
 
 class DestinationRenderNotFound : DestinationRender {
 
@@ -20,6 +20,6 @@ class DestinationRenderNotFound : DestinationRender {
         navController: NavHostController,
         navBackStackEntry: NavBackStackEntry
     ) {
-        DestinationRenderNotFoundView(destinationInfo.renderType)
+        MacaoDestinationRenderNotFoundView(destinationInfo.renderType)
     }
 }
