+++ composeApp/src/commonMain/kotlin/com.macaosoftware.component.navigationcompose.demo/marketplace/notfound/RootDestinationRenderNotFound.kt
@@ -1,10 +1,10 @@
-package com.macaosoftware.component.navigationcompose.demo.serverui.domain.error
+package com.macaosoftware.component.navigationcompose.demo.marketplace.notfound
 
 import androidx.compose.runtime.Composable
 import com.macaosoftware.component.core.DestinationInfo
 import com.macaosoftware.component.core.RootDestinationRender
 import com.macaosoftware.component.navigationcompose.demo.serverui.data.ServerUiConstants
-import com.macaosoftware.component.util.DestinationRenderNotFoundView
+import com.macaosoftware.component.util.MacaoDestinationRenderNotFoundView
 
 class RootDestinationRenderNotFound : RootDestinationRender {
 
@@ -14,6 +14,6 @@ class RootDestinationRenderNotFound : RootDestinationRender {
 
     @Composable
     override fun Content(destinationInfo: DestinationInfo) {
-        DestinationRenderNotFoundView(destinationInfo.renderType)
+        MacaoDestinationRenderNotFoundView(destinationInfo.renderType)
     }
 }
